Restructure the folder
- make MxHeader and MxFooter
- Move the appbar2 URL to .env
