+++ .env.production
@@ -1,0 +1,1 @@
+REACT_APP_PROFILE_URL=https://home.mendix.com/mxid/appbar2
