+++ .env.test
@@ -1,0 +1,1 @@
+REACT_APP_PROFILE_URL=https://home-test.mendix.com/mxid/appbar2
