+++ src/components/MxFooter/index.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const MxFooter = () => {
+    return <div>Footer</div>;
+};
+
+export default MxFooter;
