+++ src/components/MxHeader/Settings/Profile.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import { connect } from '../../context/store';
-import defaultAvatar from '../../resources/img/header/avatar-empty.png';
+import { connect } from '../../../context/store';
+import defaultAvatar from '../../../resources/img/header/avatar-empty.png';
 
 import ProfileMenu from './ProfileMenu';
