+++ src/components/MxHeader/Settings/StartButton.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { links } from '../../resources/mendix.json';
+import { links } from '../../../resources/mendix.json';
 
 const StartButton = () => (
     <a href={links.signup} className="MxHeader__start-button">
