+++ src/index.tsx
@@ -1,5 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
 
-ReactDOM.render(<App />, document.getElementById('root'));
+import MxHeader from './components/MxHeader';
+import MxFooter from './components/MxFooter';
+
+import './style/MxHeader.scss';
+import './style/MxFooter.scss';
+
+ReactDOM.render(<MxHeader />, document.getElementById('mxHeader'));
+ReactDOM.render(<MxFooter />, document.getElementById('mxFooter'));
