+++ src/style/MxFooter.scss
@@ -1,0 +1,7 @@
+@import '../../node_modules/reset-css/sass/reset';
+@import 'variables';
+@import 'helpers';
+
+* {
+    box-sizing: border-box;
+}
