+++ src/style/_variables.scss
@@ -14,7 +14,7 @@ $color-top-background: $color-white;
 $color-top-border: $color-mx-blue;
 
 $color-sub-background: #ededed;
-$color-submenu-border: #d2d2d2;
+$color-divider: #d2d2d2;
 
 $color-bottom-background: $color-bottom-blue;
 $color-bottom-font: #a4acbb;
